#15: Created projects library for generate-workflows script
