+++ scripts/generate-workflows/pkg/projects/materialize.go
@@ -1,4 +1,4 @@
-package main
+package projects
 
 import (
 	"fmt"
